+++ tests/testthat/test-precision.r
@@ -1,8 +1,11 @@
 test_that("precision", {
-    library("clock", exclude = c("calendar_narrow", "calendar_widen"))
+    library("clock", exclude = c("calendar_narrow", "calendar_precision", "calendar_widen"))
     dts <- as_datetimeoffset(c("2020", "2020-04-10", "2020-04-10T10:10"))
     expect_equal(calendar_precision(dts),
                  c("year", "day", "minute"))
+    expect_equal(calendar_precision(dts, range = TRUE),
+                 c("year", "minute"))
+
     dtn <- calendar_narrow(dts, "day")
     expect_equal(format(dtn), c("2020", "2020-04-10", "2020-04-10"))